+++ paper/paper.bib
@@ -159,7 +159,7 @@ year = {2017}
   number={1},
   pages={2171--2175},
   year={2012},
-  doi={10.5555/2503308.2503311},
+  url = {http://jmlr.org/papers/v13/fortin12a.html},
   publisher={JMLR. org}
 }
 
@@ -168,6 +168,7 @@ year = {2017}
   author={Marelli, Stefano and Sudret, Bruno},
   booktitle={Vulnerability, uncertainty, and risk: quantification, mitigation, and management},
   pages={2554--2563},
+  doi={10.1061/9780784413609.257},
   year={2014}
 }
 
