### 2022-03-18 3c6ffdc
add check for only 1 quantity of interest
