### 2022-04-20 #4131 @EricaEgg
Apologies, I am also not able to review right now.

On Wed, Apr 20, 2022 at 3:02 AM fraukewiese ***@***.***>
wrote:

> @wholmgren <https://github.com/wholmgren>, @EricaEgg
> <https://github.com/EricaEgg>, @kramea <https://github.com/kramea> - to
> make sure you have seen this request, I send it again:
> – would any of you be willing to review this submission for JOSS? We carry
> out our checklist-driven reviews here in GitHub issues and follow these
> guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The submission under review here is RHEIA: Robust design optimization of
> renewable Hydrogen and dErIved energy cArrier systems
> <https://github.com/rheia-framework/RHEIA>
>
