### 2022-04-21 #4131 @fraukewiese
@ClaraBuettner , @kramea:
– would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is [RHEIA: Robust design optimization of renewable Hydrogen and dErIved energy cArrier systems](https://github.com/rheia-framework/RHEIA)
