### 2022-05-04 #4370 @editorialbot
[REVIEW]: RHEIA: Robust design optimization of renewable Hydrogen and dErIved energy cArrier systems
**Submitting author:** @DCoppitters (<a href="http://orcid.org/0000-0001-9480-2781">Diederik Coppitters</a>)
**Repository:** https://github.com/rheia-framework/RHEIA
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
