### 2022-05-09 4cc4e3f
Fix to issue #10
