### 2022-06-08 187fe0e
Solved issue with logo not appearing
