### 2022-06-30 #4370 @DCoppitters
@andr1976 and @ClaraBuettner : Thanks a lot for your review, time and efforts!

@fraukewiese : Below the required information:
- v1.0.0
- 10.5281/zenodo.6782705

Thanks a lot for your time and efforts!
