+++ paper/paper.md
@@ -53,12 +53,12 @@ hydrogen can be converted back into electricity (power-to-power),
 it can be used to produce low-carbon fuels (power-to-fuel),
 and it can be used to fuel hydrogen vehicles (power-to-mobility).
 The performance of these hydrogen-based energy systems is subject to uncertainties, 
-such as the uncertainty on the solar irradiance, the energy consumption of hydrogen-powered buses and the price of grid electricity.
+such as the uncertainty on the solar irradiance, the energy consumption of hydrogen-powered buses, and the price of grid electricity.
 Disregarding these uncertainties in the design process can result in a drastic mismatch between simulated and real-world performance, 
 and thus lead to a *kill-by-randomness* of the system.
 The *Robust design optimization of renewable Hydrogen and dErIved energy cArrier systems* (RHEIA) framework provides a robust design optimization pipeline 
 that considers real-world uncertainties and yields robust designs, i.e., designs with a performance less sensitive to these uncertainties.
-Moreover, RHEIA includes models to evaluate hydrogen's techno-economic and environmental performance in a power-to-fuel, power-to-power and power-to-mobility context.
+Moreover, RHEIA includes models to evaluate hydrogen's techno-economic and environmental performance in a power-to-fuel, power-to-power, and power-to-mobility context.
 When combined, RHEIA unlocks the robust designs for hydrogen-based energy systems.
 As RHEIA considers the system models as a black box, the framework can be applied to existing open-source and closed-source models.
 To illustrate, an interface with the [EnergyPLAN](https://www.energyplan.eu/) software is included in the framework. 
@@ -68,22 +68,22 @@ To illustrate, an interface with the [EnergyPLAN](https://www.energyplan.eu/) so
 
 Incorporating hydrogen is still an anomaly in design optimization studies of renewable energy systems [@Eriksson2017]. 
 Moreover, the optimization is often performed under the assumption of deterministic parameters (i.e., fixed, free from inherent variation).
-Considering fixed values for model parameters in design optimization yields designs that might be sensitive -- the real issue is that we cannot know-- to real-world uncertainties
+Considering fixed values for model parameters in design optimization yields designs that might be sensitive -- the real issue is that we cannot know -- to real-world uncertainties
 and results in a drastic mismatch between simulated and actual performances.
-In fields different from energy systems, e.g., structural mechanics, aerospace and automobile applications, 
-Robust Design Optimization (RDO) yielded robust designs by minimizing the variance on the performance [@orosz2020robust].
+In fields different from energy systems, e.g., structural mechanics, aerospace, and automobile applications, 
+Robust Design Optimization (RDO) yields robust designs by minimizing the variance on the performance [@orosz2020robust].
 Consequently, alternative design solutions were proposed that provide the least sensitive performance to the random environment.
 To ensure the computational tractability of RDO, surrogate modelling techniques achieve a promising computational efficiency
-to quantify the mean and variance on the performance. Nevertheless, applications of such surrogate-assisted robust design optimization techniques are limited [@Chatterjee2017].
+to quantify the mean and variance of the performance. Nevertheless, applications of such surrogate-assisted robust design optimization techniques are limited [@Chatterjee2017].
 To fill these research gaps, RHEIA provides a multi-objective RDO algorithm,
 for which the uncertainty quantification is performed through a Polynomial Chaos Expansion (PCE) surrogate modelling technique.
-In addition, RHEIA includes Python-based models for relevant valorization pathways of hydrogen: power-to-fuel, power-to-power and power-to-mobility.
[…]
