### 2022-07-05 #4370 @danielskatz
@DCoppitters - I'm the AEiC this week who will complete processing of this paper.  I've suggested some small changes to the paper in https://github.com/rheia-framework/RHEIA/pull/14 - please merge this or let me know what you disagree with, then we can proceed.
